+++ src/Git_Updater/Base.php
@@ -517,7 +517,8 @@ class Base {
 		$new_source = $this->fix_misnamed_directory( $new_source, $remote_source, $upgrader_object, $slug );
 
 		if ( $source !== $new_source ) {
-			$result = $this->move_dir( $source, $new_source );
+			new Shim();
+			$result = move_dir( $source, $new_source );
 			if ( \is_wp_error( $result ) ) {
 				return $result;
 			}