+++ src/Git_Updater/Shim.php
@@ -14,6 +14,7 @@ namespace Fragen\Git_Updater;
  * Class Shim
  *
  * Provides PHP 5.6 compatible shims.
+ * Loads WP 6.1 modified functions from Rollback.
  */
 class Shim {
 
@@ -39,4 +40,183 @@ class Shim {
 			}
 		}
 	}
+}
+
+/**
+ * Moves a directory from one location to another via the rename() PHP function.
+ * If the renaming failed, falls back to copy_dir().
+ *
+ * Assumes that WP_Filesystem() has already been called and setup.
+ *
+ * @since 6.1.0
+ *
+ * @global WP_Filesystem_Base $wp_filesystem WordPress filesystem subclass.
+ *
+ * @param string $from        Source directory.
+ * @param string $to          Destination directory.
+ * @return true|WP_Error True on success, WP_Error on failure.
+ */
+function move_dir( $from, $to ) {
+	global $wp_filesystem;
+
+	$result = false;
+
+	/*
+	 * Skip the rename() call on VirtualBox environments.
+	 * There are some known issues where rename() can fail on shared folders
+	 * without reporting an error properly.
+	 *
+	 * More details:
+	 * https://www.virtualbox.org/ticket/8761#comment:24
+	 * https://www.virtualbox.org/ticket/17971
+	 */
+	if ( 'direct' === $wp_filesystem->method && ! is_virtualbox() ) {
+		$wp_filesystem->rmdir( $to );
+
+		$result = @rename( $from, $to );
+	}
+
+	// Non-direct filesystems use some version of rename without a fallback.
+	if ( 'direct' !== $wp_filesystem->method ) {
+		$result = $wp_filesystem->move( $from, $to );
+	}
+
+	if ( ! $result ) {
+		if ( ! $wp_filesystem->is_dir( $to ) ) {
+			if ( ! $wp_filesystem->mkdir( $to, FS_CHMOD_DIR ) ) {
+				return new \WP_Error( 'mkdir_failed_move_dir', __( 'Could not create directory.' ), $to );
+			}
+		}
+
+			$result = copy_dir( $from, $to );
+	}
+
+	return $result;
+}
+
+/**
+ * Recursive file/directory copy and delete.
+ *
+ * Functions more like `rename()` in that the $source is deleted after copying.
+ *
+ * More versatile than WP Core `copy_dir()`.
+ *
+ * @param string $source      File path of source.
+ * @param string $destination File path of destination.
+ *
+ * @return bool|\WP_Error True for success, \WP_Error for failure.
+ */
+function copy_dir( $source, $destination ) {
+	// phpcs:ignore WordPress.CodeAnalysis.AssignmentInCondition.Found, Squiz.PHP.DisallowMultipleAssignments.FoundInControlStructure
+	if ( $dir = opendir( $source ) ) {
+		if ( ! file_exists( $destination ) ) {
+			mkdir( $destination );
+		}
+		$source = untrailingslashit( $source );
+		// phpcs:ignore WordPress.CodeAnalysis.AssignmentInCondition.FoundInWhileCondition
+		while ( false !== ( $file = readdir( $dir ) ) ) {
+			if ( ( '.' !== $file ) && ( '..' !== $file ) && "{$source}/{$file}" !== $destination ) {
+				if ( is_dir( "{$source}/{$file}" ) ) {
+					move_dir( "{$source}/{$file}", "{$destination}/{$file}" );
+				} else {
+					copy( "{$source}/{$file}", "{$destination}/{$file}" );
+					unlink( "{$source}/{$file}" );
+				}
+			}
+		}
+		$iterator = new \FilesystemIterator( $source );
+		if ( ! $iterator->valid() ) { // True if directory is empty.
+			rmdir( $source );
+		}
+		closedir( $dir );
+
+		return true;
+	}
+
+	return new \WP_Error( 'recursive_copy_delete_failed', __( 'Could not move directory using `recursive_copy_delete`.' ), [ $source, $destination ] );
+}
+
+/**
+ * Attempt to detect a VirtualBox environment.
+ *
+ * This attempts all known methods of detecting VirtualBox.
+ *
+ * @global $wp_filesystem The filesystem.
+ *
+ * @since 6.1.0
+ *
+ * @return bool Whether or not VirtualBox was detected.
+ */
+function is_virtualbox() {
+	global $wp_filesystem;
+	static $is_virtualbox;
+
+	if ( ! defined( 'WP_RUN_CORE_TESTS' ) && null !== $is_virtualbox ) {
+		return $is_virtualbox;
+	}
+
+	/*
+	 * Filters whether the current environment uses VirtualBox.
+	 *
+	 * @since 6.1.0
+	 *
+	 * @param bool $is_virtualbox Whether the current environment uses VirtualBox.
+	 *                            Default: false.
+	 */
+	if ( apply_filters( 'is_virtualbox', false ) ) {
+		$is_virtualbox = true;
+		return $is_virtualbox;
+	}
+
+	// Detection via Composer.
+	if ( function_exists( 'getenv' ) && 'virtualbox' === getenv( 'COMPOSER_RUNTIME_ENV' ) ) {
+		$is_virtualbox = true;
+		return $is_virtualbox;
+	}
+
+	$virtualbox_unames = [ 'vvv' ];
+
+	// Detection via `php_uname()`.
+	if ( function_exists( 'php_uname' ) && in_array( php_uname( 'n' ), $virtualbox_unames, true ) ) {
+		$is_virtualbox = true;
+		return $is_virtualbox;
+	}
+
+	/*
+	 * Vagrant can use alternative providers.
+	 * This isn't reliable without some additional check(s).
+	 */
+	$virtualbox_usernames = [ 'vagrant' ];
+
+	// Detection via user name with POSIX.
+	if ( function_exists( 'posix_getpwuid' ) && function_exists( 'posix_geteuid' ) ) {
+		$user = posix_getpwuid( posix_geteuid() );
+		if ( $user && in_array( $user['name'], $virtualbox_usernames, true ) ) {
+			$is_virtualbox = true;
+			return $is_virtualbox;
+		}
+	}
+
+	// Initialize the filesystem if not set.
+	if ( ! $wp_filesystem ) {
+		require_once ABSPATH . '/wp-admin/includes/file.php';
+		WP_Filesystem();
+	}
+
+	// Detection via file owner.
+	if ( in_array( $wp_filesystem->owner( __FILE__ ), $virtualbox_usernames, true ) ) {
+		$is_virtualbox = true;
+		return $is_virtualbox;
+	}
+
+	// Detection via file group.
+	if ( in_array( $wp_filesystem->group( __FILE__ ), $virtualbox_usernames, true ) ) {
+		$is_virtualbox = true;
+		return $is_virtualbox;
+	}
+
+	// Give up.
+	$is_virtualbox = false;
+
+	return $is_virtualbox;
 }